+++ src/WebLinter/Linters/CoffeeLinter.cs
@@ -5,39 +5,18 @@ namespace WebLinter
 {
     internal class CoffeeLinter : LinterBase
     {
-        private static Regex _rx = new Regex(@"\[stdin\]:(?<line>[0-9]+):(?<column>[0-9]+): error: (?<message>.+)", RegexOptions.Compiled | RegexOptions.IgnoreCase);
+        private static Regex _rx = new Regex(@"^(?<file>[^,]+),(?<line>[0-9]+),(?<column>[0-9]+)?,.+,(?<message>[^,]+)$", RegexOptions.Compiled | RegexOptions.IgnoreCase | RegexOptions.Multiline);
 
-        public CoffeeLinter(ISettings settings) : base(settings)
-        { }
-
-        public override string Name
+        public CoffeeLinter(ISettings settings) : base(settings, _rx)
         {
-            get { return "CoffeeLint"; }
+            Name = "CoffeeLint";
+            ConfigFileName = "coffeelint.json";
+            IsEnabled = Settings.CoffeeLintEnable;
         }
 
-        public override bool IsEnabled
+        protected override string GetArguments(FileInfo[] files)
         {
-            get { return Settings.CoffeeLintEnable; }
-        }
-
-        protected override LintingResult Lint(FileInfo file)
-        {
-            string output, error;
-            RunProcess(file, "coffeelint.cmd", out output, out error);
-
-            if (!string.IsNullOrEmpty(output))
-            {
-                foreach (Match match in _rx.Matches(output))
-                {
-                    AddError(file, match, Settings.CoffeeLintAsErrors);
-                }
-            }
-            else if (!string.IsNullOrEmpty(error))
-            {
-                Result.Errors.Add(new LintingError(file.FullName, error));
-            }
-
-            return Result;
+            return "--reporter=csv";
         }
     }
 }