+++ src/WebLinterTest/CoffeelintTest.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using WebLinter;
 
@@ -10,15 +10,24 @@ namespace WebLinterTest
         [TestMethod, TestCategory("CoffeeLint")]
         public void Standard()
         {
-            var result = LinterFactory.Lint("../../artifacts/coffeelint/a.coffee", Settings.Instance);
+            var result = LinterFactory.Lint(Settings.Instance, "../../artifacts/coffeelint/a.coffee");
             Assert.IsTrue(result.HasErrors);
-            Assert.AreEqual(1, result.Errors.Count, $"Found {result.Errors.Count} errors");
+            Assert.IsFalse(string.IsNullOrEmpty(result.Errors.First().FileName), "File name is empty");
+            Assert.AreEqual(2, result.Errors.Count, $"Found {result.Errors.Count} errors");
         }
+        [TestMethod, TestCategory("CoffeeLint")]
+        public void Multiple()
+        {
+            var result = LinterFactory.Lint(Settings.Instance, "../../artifacts/coffeelint/a.coffee", "../../artifacts/coffeelint/b.coffee");
+            Assert.IsTrue(result.HasErrors);
+            Assert.AreEqual(4, result.Errors.Count, $"Found {result.Errors.Count} errors");
+        }
+
 
         [TestMethod, TestCategory("CoffeeLint")]
         public void FileDontExist()
         {
-            var result = LinterFactory.Lint("../../artifacts/coffeelint/doesntexist.coffee", Settings.Instance);
+            var result = LinterFactory.Lint(Settings.Instance, "../../artifacts/coffeelint/doesntexist.coffee");
             Assert.IsTrue(result.HasErrors);
         }
     }